+++ LiteLoader/include/llapi/mc/CommandOutput.hpp
@@ -25,7 +25,7 @@
  */
 class CommandOutput {
 
-#define AFTER_EXTRA // Add members here
+#define AFTER_EXTRA
 public:
     /**
      * @brief Output a normal message.
@@ -68,7 +68,8 @@ public:
         PluginOwnData::set<std::string>(POD_COMMANDOUTPUT_LANGCODE, code);
     }
 
-    template <typename... Args> inline void trAddMessage(const std::string& format, Args&&... args) {
+    template <typename... Args>
+    inline void trAddMessage(const std::string& format, Args&&... args) {
         if (PluginOwnData::has(POD_COMMANDOUTPUT_LANGCODE))
             this->addMessage(
                 trl(PluginOwnData::get<std::string>(POD_COMMANDOUTPUT_LANGCODE), format, std::forward<Args>(args)...));
@@ -76,7 +77,8 @@ public:
             this->addMessage(tr(format, std::forward<Args>(args)...));
     }
 
-    template <typename... Args> inline void trSuccess(const std::string& format, Args&&... args) {
+    template <typename... Args>
+    inline void trSuccess(const std::string& format, Args&&... args) {
         if (PluginOwnData::has(POD_COMMANDOUTPUT_LANGCODE))
             this->success(
                 trl(PluginOwnData::get<std::string>(POD_COMMANDOUTPUT_LANGCODE), format, std::forward<Args>(args)...));
@@ -84,7 +86,8 @@ public:
             this->success(tr(format, std::forward<Args>(args)...));
     }
 
-    template <typename... Args> inline void trError(const std::string& format, Args&&... args) {
+    template <typename... Args>
+    inline void trError(const std::string& format, Args&&... args) {
         if (PluginOwnData::has(POD_COMMANDOUTPUT_LANGCODE))
             this->error(
                 trl(PluginOwnData::get<std::string>(POD_COMMANDOUTPUT_LANGCODE), format, std::forward<Args>(args)...));
@@ -95,117 +98,109 @@ public:
 #undef AFTER_EXTRA
 #ifndef DISABLE_CONSTRUCTOR_PREVENTION_COMMANDOUTPUT
 public:
-    class CommandOutput& operator=(class CommandOutput const&) = delete;
+    class CommandOutput& operator=(class CommandOutput const &) = delete;
     CommandOutput() = delete;
 #endif
 
 public:
     /**
      * @symbol ??0CommandOutput@@QEAA@W4CommandOutputType@@@Z
-     * @hash   -1906512777
+     * @hash   1635568679
      */
     MCAPI CommandOutput(enum class CommandOutputType);
     /**
      * @symbol ??0CommandOutput@@QEAA@AEBV0@@Z
-     * @hash   -1418144527
+     * @hash   2123921553
      */
-    MCAPI CommandOutput(class CommandOutput const&);
+    MCAPI CommandOutput(class CommandOutput const &);
     /**
      * @symbol ?addToResultList@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@0@Z
-     * @hash   1625676890
+     * @hash   872837178
      */
-    MCAPI void addToResultList(std::string const&, std::string const&);
+    MCAPI void addToResultList(std::string const &, std::string const &);
     /**
-     * @symbol
-     * ?addToResultList@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBVActor@@@Z
-     * @hash   -2019758587
+     * @symbol ?addToResultList@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBVActor@@@Z
+     * @hash   1522368997
      */
-    MCAPI void addToResultList(std::string const&, class Actor const&);
+    MCAPI void addToResultList(std::string const &, class Actor const &);
     /**
      * @symbol ?empty@CommandOutput@@QEBA_NXZ
-     * @hash   2114801942
+     * @hash   1361962230
      */
     MCAPI bool empty() const;
     /**
-     * @symbol
-     * ?error@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
-     * @hash   1420096547
+     * @symbol ?error@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
+     * @hash   667256835
      */
-    MCAPI void error(std::string const&, std::vector<class CommandOutputParameter> const&);
+    MCAPI void error(std::string const &, std::vector<class CommandOutputParameter> const &);
     /**
-     * @symbol
-     * ?forceOutput@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
-     * @hash   1031751947
+     * @symbol ?forceOutput@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
+     * @hash   278912235
      */
-    MCAPI void forceOutput(std::string const&, std::vector<class CommandOutputParameter> const&);
+    MCAPI void forceOutput(std::string const &, std::vector<class CommandOutputParameter> const &);
     /**
      * @symbol ?getData@CommandOutput@@QEBAAEBVCommandPropertyBag@@XZ
-     * @hash   -1317558974
+     * @hash   -1338854734
      */
-    MCAPI class CommandPropertyBag const& getData() const;
+    MCAPI class CommandPropertyBag const & getData() const;
     /**
-     * @symbol
-     * ?getMessages@CommandOutput@@QEBAAEBV?$vector@VCommandOutputMessage@@V?$allocator@VCommandOutputMessage@@@std@@@std@@XZ
-     * @hash   -788988960
+     * @symbol ?getMessages@CommandOutput@@QEBAAEBV?$vector@VCommandOutputMessage@@V?$allocator@VCommandOutputMessage@@@std@@@std@@XZ
+     * @hash   405161920
      */
-    MCAPI std::vector<class CommandOutputMessage> const& getMessages() const;
+    MCAPI std::vector<class CommandOutputMessage> const & getMessages() const;
     /**
      * @symbol ?getSuccessCount@CommandOutput@@QEBAHXZ
-     * @hash   1360894636
+     * @hash   509909916
      */
     MCAPI int getSuccessCount() const;
     /**
      * @symbol ?getType@CommandOutput@@QEBA?AW4CommandOutputType@@XZ
-     * @hash   -1159705490
+     * @hash   -1079950178
      */
     MCAPI enum class CommandOutputType getType() const;
     /**
      * @symbol ?hasPlayerText@CommandOutput@@QEBA_NXZ
-     * @hash   1611387254
+     * @hash   858547542
      */
     MCAPI bool hasPlayerText() const;
     /**
-     * @symbol
-     * ?load@CommandOutput@@QEAAXW4CommandOutputType@@H$$QEAV?$vector@VCommandOutputMessage@@V?$allocator@VCommandOutputMessage@@@std@@@std@@$$QEAV?$unique_ptr@VCommandPropertyBag@@U?$default_delete@VCommandPropertyBag@@@std@@@4@@Z
-     * @hash   847267746
+     * @symbol ?load@CommandOutput@@QEAAXW4CommandOutputType@@H$$QEAV?$vector@VCommandOutputMessage@@V?$allocator@VCommandOutputMessage@@@std@@@std@@$$QEAV?$unique_ptr@VCommandPropertyBag@@U?$default_delete@VCommandPropertyBag@@@std@@@4@@Z
+     * @hash   94428034
      */
-    MCAPI void load(enum class CommandOutputType, int, std::vector<class CommandOutputMessage>&&,
-                    std::unique_ptr<class CommandPropertyBag>&&);
+    MCAPI void load(enum class CommandOutputType, int, std::vector<class CommandOutputMessage> &&, std::unique_ptr<class CommandPropertyBag> &&);
     /**
      * @symbol ?setHasPlayerText@CommandOutput@@QEAAXXZ
-     * @hash   1334051535
+     * @hash   581227199
      */
     MCAPI void setHasPlayerText();
     /**
-     * @symbol
-     * ?success@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
-     * @hash   1780476233
+     * @symbol ?success@CommandOutput@@QEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@@Z
+     * @hash   1027651897
      */
-    MCAPI void success(std::string const&, std::vector<class CommandOutputParameter> const&);
+    MCAPI void success(std::string const &, std::vector<class CommandOutputParameter> const &);
     /**
      * @symbol ?success@CommandOutput@@QEAAXXZ
-     * @hash   163386203
+     * @hash   -589438133
      */
     MCAPI void success();
     /**
      * @symbol ?wantsData@CommandOutput@@QEBA_NXZ
-     * @hash   -1266489514
+     * @hash   -2019313850
      */
     MCAPI bool wantsData() const;
     /**
      * @symbol ??1CommandOutput@@QEAA@XZ
-     * @hash   56722184
+     * @hash   41684456
      */
     MCAPI ~CommandOutput();
 
-    // private:
+//private:
     /**
-     * @symbol
-     * ?addMessage@CommandOutput@@AEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@W4CommandOutputMessageType@@@Z
-     * @hash   -2124484666
+     * @symbol ?addMessage@CommandOutput@@AEAAXAEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@AEBV?$vector@VCommandOutputParameter@@V?$allocator@VCommandOutputParameter@@@std@@@3@W4CommandOutputMessageType@@@Z
+     * @hash   1417627542
      */
-    MCAPI void addMessage(std::string const&, std::vector<class CommandOutputParameter> const&,
-                          enum class CommandOutputMessageType);
+    MCAPI void addMessage(std::string const &, std::vector<class CommandOutputParameter> const &, enum class CommandOutputMessageType);
 
 private:
+
 };