+++ LiteLoader/src/llapi/mc/ActorAPI.cpp
@@ -21,13 +21,13 @@
 #include "llapi/mc/TeleportTarget.hpp"
 #include "llapi/mc/UserEntityIdentifierComponent.hpp"
 #include "llapi/mc/OnFireSystem.hpp"
-#include "llapi/mc/TeleportRotationData.hpp"
 #include "llapi/mc/ClipDefaults.hpp"
 #include "llapi/mc/ItemStack.hpp"
 #include "llapi/mc/ActorDefinitionIdentifier.hpp"
 #include "llapi/mc/ActorDamageSource.hpp"
 #include "llapi/mc/Vec2.hpp"
 #include "llapi/mc/AABB.hpp"
+#include "llapi/mc/RotationCommandUtils.hpp"
 
 class UserEntityIdentifierComponent;
 
@@ -66,7 +66,7 @@ bool Actor::isItemActor() const {
 }
 
 bool Actor::isOnGround() const {
-    return (dAccess<bool, 448>(this)); // IDA DirectActorProxyImpl<IMobMovementProxy>::isOnGround
+    return (dAccess<bool, 440>(this)); // IDA DirectActorProxyImpl<IMobMovementProxy>::isOnGround
 }
 
 std::string Actor::getTypeName() const {
@@ -102,21 +102,29 @@ ActorUniqueID Actor::getActorUniqueId() const {
     }
 }
 
-static_assert(sizeof(TeleportRotationData) == 32);
+static_assert(sizeof(RotationCommandUtils::RotationData) == 32);
+
 bool Actor::teleport(Vec3 to, int dimID) {
     if (!this->isAlive())
         return false;
     char mem[48];
-    auto computeTarget = (TeleportTarget * (*)(void*, class Actor&, class Vec3, class Vec3*, class AutomaticID<class Dimension, int>, std::optional<TeleportRotationData> const&, int))(&TeleportCommand::computeTarget);
-    auto target = computeTarget(mem, *this, to, nullptr, dimID, TeleportRotationData{getRotation().x, getRotation().y, {}}, 15);
+    auto computeTarget =
+        (TeleportTarget * (*)(void*, class Actor&, class Vec3, class Vec3*, class AutomaticID<class Dimension, int>,
+                              std::optional<RotationCommandUtils::RotationData> const&,
+                              int))(&TeleportCommand::computeTarget);
+    auto target = computeTarget(mem, *this, to, nullptr, dimID,
+                                RotationCommandUtils::RotationData{getRotation().x, getRotation().y, {}}, 15);
     TeleportCommand::applyTarget(*this, *target, false);
     return true;
 }
 
 bool Actor::teleport(Vec3 to, int dimID, float x, float y) {
     char mem[48];
-    auto computeTarget = (TeleportTarget * (*)(void*, class Actor&, class Vec3, class Vec3*, class AutomaticID<class Dimension, int>, std::optional<TeleportRotationData> const&, int))(&TeleportCommand::computeTarget);
-    auto target = computeTarget(mem, *this, to, nullptr, dimID, TeleportRotationData{x, y, {}}, 15);
+    auto computeTarget =
+        (TeleportTarget * (*)(void*, class Actor&, class Vec3, class Vec3*, class AutomaticID<class Dimension, int>,
+                              std::optional<RotationCommandUtils::RotationData> const&,
+                              int))(&TeleportCommand::computeTarget);
+    auto target = computeTarget(mem, *this, to, nullptr, dimID, RotationCommandUtils::RotationData{x, y, {}}, 15);
     TeleportCommand::applyTarget(*this, *target, false);
     return true;
 }