+++ ScriptEngine/src/api/PlayerAPI.cpp
@@ -15,6 +15,8 @@
 #include "engine/EngineOwnData.h"
 #include "engine/GlobalShareData.h"
 #include "main/EconomicSystem.h"
+#include <llapi/mc/Attribute.hpp>
+#include <llapi/mc/AttributeInstance.hpp>
 #include <llapi/mc/Player.hpp>
 #include <llapi/mc/NetworkIdentifier.hpp>
 #include <llapi/mc/Actor.hpp>
@@ -27,6 +29,7 @@
 #include <llapi/mc/CompoundTag.hpp>
 #include <llapi/mc/SynchedActorData.hpp>
 #include <llapi/mc/SimulatedPlayer.hpp>
+#include <llapi/mc/SharedAttributes.hpp>
 #include <llapi/mc/BlockSource.hpp>
 #include <llapi/mc/Command.hpp>
 #include <llapi/mc/SynchedActorDataEntityWrapper.hpp>
@@ -121,13 +124,17 @@ ClassDefine<PlayerClass> PlayerClassBuilder =
         .instanceFunction("crash", &PlayerClass::crash)
         .instanceFunction("hurt", &PlayerClass::hurt)
         .instanceFunction("heal", &PlayerClass::heal)
+        .instanceFunction("setHealth", &PlayerClass::setHealth)
+        .instanceFunction("setMaxHealth", &PlayerClass::setMaxHealth)
+        .instanceFunction("setHungry", &PlayerClass::setHungry)
         .instanceFunction("refreshChunks", &PlayerClass::refreshChunks)
         .instanceFunction("giveItem", &PlayerClass::giveItem)
         .instanceFunction("clearItem", &PlayerClass::clearItem)
         .instanceFunction("isSprinting", &PlayerClass::isSprinting)
         .instanceFunction("setSprinting", &PlayerClass::setSprinting)
         .instanceFunction("sendToast", &PlayerClass::sendToast)
-        .instanceFunction("distanceToPos", &PlayerClass::distanceToPos)
+        .instanceFunction("distanceTo", &PlayerClass::distanceTo)
+        .instanceFunction("distanceToSqr", &PlayerClass::distanceToSqr)
 
         .instanceFunction("getBlockStandingOn", &PlayerClass::getBlockStandingOn)
         .instanceFunction("getDevice", &PlayerClass::getDevice)
@@ -222,6 +229,7 @@ ClassDefine<PlayerClass> PlayerClassBuilder =
         .instanceFunction("removeItem", &PlayerClass::removeItem)
         .instanceFunction("getAllItems", &PlayerClass::getAllItems)
         .instanceFunction("removeScore", &PlayerClass::deleteScore)
+        .instanceFunction("distanceToPos", &PlayerClass::distanceTo)
         .build();
 
 
@@ -2060,6 +2068,60 @@ Local<Value> PlayerClass::heal(const Arguments& args) {
     CATCH("Fail in heal!");
 }
 
+Local<Value> PlayerClass::setHealth(const Arguments& args) {
+    CHECK_ARGS_COUNT(args, 1);
+    CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
+
+    try {
+        Player* player = get();
+        if (!player)
+            return Local<Value>();
+
+        AttributeInstance* healthAttribute = player->getMutableAttribute(Global<SharedAttributes>->HEALTH);
+
+        healthAttribute->setCurrentValue(args[0].asNumber().toFloat());
+
+        return Boolean::newBoolean(true);
+    }
+    CATCH("Fail in setHealth!");
+}
+
+Local<Value> PlayerClass::setMaxHealth(const Arguments& args) {
+    CHECK_ARGS_COUNT(args, 1);
+    CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
+
+    try {
+        Player* player = get();
+        if (!player)
+            return Local<Value>();
+
+        AttributeInstance* healthAttribute = player->getMutableAttribute(Global<SharedAttributes>->HEALTH);
+
+        healthAttribute->setMaxValue(args[0].asNumber().toFloat());
+
+        return Boolean::newBoolean(true);
+    }
+    CATCH("Fail in setMaxHealth!");
+}
+
+Local<Value> PlayerClass::setHungry(const Arguments& args) {
+    CHECK_ARGS_COUNT(args, 1);
+    CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
+
+    try {
+        Player* player = get();
+        if (!player)
+            return Local<Value>();
+
+        AttributeInstance* healthAttribute = player->getMutableAttribute(player->HUNGER);
+
+        healthAttribute->setCurrentValue(args[0].asNumber().toFloat());
+
+        return Boolean::newBoolean(true);
+    }
+    CATCH("Fail in setHungry!");
+}
+
 Local<Value> PlayerClass::setFire(const Arguments& args) {
     CHECK_ARGS_COUNT(args, 2);
     CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
@@ -2131,7 +2193,11 @@ Local<Value> PlayerClass::giveItem(const Arguments& args) {
         auto item = ItemClass::extract(args[0]);
         if (!item)
             return Local<Value>(); // Null
-
+        if (args.size() >= 2)
+        {
+            CHECK_ARG_TYPE(args[1], ValueKind::kNumber);
+            return Boolean::newBoolean(player->giveItem(item, args[1].toInt()));
+        }
         return Boolean::newBoolean(player->giveItem(item));
     }
     CATCH("Fail in giveItem!");
@@ -2571,7 +2637,7 @@ Local<Value> PlayerClass::sendToast(const Arguments& args) {
     CATCH("Fail in sendToast!");
 }
 
-Local<Value> PlayerClass::distanceToPos(const Arguments& args) {
+Local<Value> PlayerClass::distanceTo(const Arguments& args) {
     CHECK_ARGS_COUNT(args, 1);
     if (args.size() == 4) {
         CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
@@ -2603,6 +2669,17 @@ Local<Value> PlayerClass::distanceToPos(const Arguments& args) {
                 else {
                     pos = *posObj;
                 }
+            } else if (IsInstanceOf<PlayerClass>(args[0]) || IsInstanceOf<EntityClass>(args[0])) {
+                // Player or Entity
+
+                Actor* targetActor = EntityClass::extract(args[0]);
+
+                Vec3 targetActorPos = targetActor->getPosition();
+
+                pos.x = targetActorPos.x;
+                pos.y = targetActorPos.y;
+                pos.z = targetActorPos.z;
+                pos.dim = targetActor->getDimensionId();
             } else {
                 LOG_WRONG_ARG_TYPE();
                 return Local<Value>();
@@ -2627,11 +2704,79 @@ Local<Value> PlayerClass::distanceToPos(const Arguments& args) {
         if (!player)
             return Local<Value>();
 
-        if ((int)player->getDimensionId() != pos.dim) {
-            return Local<Value>();
-        } else {
-            return Number::newNumber(player->distanceTo(pos.getVec3()));
-        }
+        return Number::newNumber(player->distanceTo(pos.getVec3()));
     }
-    CATCH("Fail in distanceToPos!")
+    CATCH("Fail in distanceTo!")
+}
+
+Local<Value> PlayerClass::distanceToSqr(const Arguments& args) {
+    CHECK_ARGS_COUNT(args, 1);
+    if (args.size() == 4) {
+        CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
+        CHECK_ARG_TYPE(args[1], ValueKind::kNumber);
+        CHECK_ARG_TYPE(args[2], ValueKind::kNumber);
+        CHECK_ARG_TYPE(args[3], ValueKind::kNumber);
+    }
+
+    try {
+        FloatVec4 pos;
+
+        if (args.size() == 1) {
+            if (IsInstanceOf<IntPos>(args[0])) {
+                // IntPos
+                IntPos* posObj = IntPos::extractPos(args[0]);
+                if (posObj->dim < 0)
+                    return Local<Value>();
+                else {
+                    pos.x = posObj->x;
+                    pos.y = posObj->y;
+                    pos.z = posObj->z;
+                    pos.dim = posObj->dim;
+                }
+            } else if (IsInstanceOf<FloatPos>(args[0])) {
+                // FloatPos
+                FloatPos* posObj = FloatPos::extractPos(args[0]);
+                if (posObj->dim < 0)
+                    return Local<Value>();
+                else {
+                    pos = *posObj;
+                }
+            } else if (IsInstanceOf<PlayerClass>(args[0]) || IsInstanceOf<EntityClass>(args[0])) {
+                // Player or Entity
+
+                Actor* targetActor = EntityClass::extract(args[0]);
+
+                Vec3 targetActorPos = targetActor->getPosition();
+
+                pos.x = targetActorPos.x;
+                pos.y = targetActorPos.y;
+                pos.z = targetActorPos.z;
+                pos.dim = targetActor->getDimensionId();
+            } else {
+                LOG_WRONG_ARG_TYPE();
+                return Local<Value>();
+            }
+        } else if (args.size() == 4) {
+            // number pos
+            CHECK_ARG_TYPE(args[0], ValueKind::kNumber);
+            CHECK_ARG_TYPE(args[1], ValueKind::kNumber);
+            CHECK_ARG_TYPE(args[2], ValueKind::kNumber);
+            CHECK_ARG_TYPE(args[3], ValueKind::kNumber);
+
+            pos.x = args[0].asNumber().toFloat();
+            pos.y = args[1].asNumber().toFloat();
+            pos.z = args[2].asNumber().toFloat();
+            pos.dim = args[3].toInt();
+        } else {
+            LOG_WRONG_ARGS_COUNT();
+            return Local<Value>();
+        }
+
+        Player* player = get();
+        if (!player)
+            return Local<Value>();
+
+        return Number::newNumber(player->distanceToSqr(pos.getVec3()));
+    }
+    CATCH("Fail in distanceToSqr!")
 }